+++ cfdm/data/raggedindexedarray.py
@@ -26,7 +26,7 @@ dimension belongs to.
 
 :Parameters:
 
-    compressed_array: numpy array-like or subclass of Array
+    compressed_array: `Data`
         The compressed array.
 
     shape: `tuple`
@@ -43,11 +43,11 @@ dimension belongs to.
         corresponding to a CF-netCDF index variable.
 
         '''
-        if not isinstance(compressed_array, abstract.Array):
-            if not isinstance(compressed_array, numpy.ndarray):
-                compressed_array = numpy.asanyarray(compressed_array)
-                
-            compressed_array = NumpyArray(compressed_array)
+#        if not isinstance(compressed_array, abstract.Array):
+#            if not isinstance(compressed_array, numpy.ndarray):
+#                compressed_array = numpy.asanyarray(compressed_array)
+#                
+#            compressed_array = NumpyArray(compressed_array)
 
         super().__init__(compressed_array=compressed_array,
                          shape=shape, size=size, ndim=ndim,