+++ cfdm/data/raggedindexedcontiguousarray.py
@@ -31,7 +31,7 @@ variable" that specifies the feature that each profile belongs to.
 
 :Parameters:
 
-    compressed_array: numpy array-like or subclass of Array
+    compressed_array: `Data`
         The compressed array.
 
     shape: `tuple`
@@ -52,11 +52,11 @@ variable" that specifies the feature that each profile belongs to.
         corresponding to a CF-netCDF CF-netCDF index variable.
 
         '''
-        if not isinstance(compressed_array, abstract.Array):
-            if not isinstance(compressed_array, numpy.ndarray):
-                compressed_array = numpy.asanyarray(compressed_array)
-                
-            compressed_array = NumpyArray(compressed_array)
+#        if not isinstance(compressed_array, abstract.Array):
+#            if not isinstance(compressed_array, numpy.ndarray):
+#                compressed_array = numpy.asanyarray(compressed_array)
+#                
+#            compressed_array = NumpyArray(compressed_array)
 
         super().__init__(compressed_array=compressed_array,
                          shape=shape, size=size, ndim=ndim,